EVENT SYSTEM - Player Controller
Modified the `PlayerController.cs` script to expose a `public` method to update internal values, which is then triggered by a `Vector2GameEventListener` on the prefab itself!

This removes a lot of code from the `PlayerController` _and_ allows us to make mappings like this in the Inspector, rather than dealing with code.